chore(iac): move iac snippets into the parent folder. This enables functional tests to use them with no issue
