+++ tests/conftest.py
@@ -31,6 +31,39 @@ skipwindows = pytest.mark.skipif(
 )
 
 
+_IAC_SINGLE_VULNERABILITY = """
+resource "aws_alb_listener" "bad_example" {
+  protocol = "HTTP"
+}
+"""
+
+
+_IAC_MULTIPLE_VULNERABILITIES = """
+resource "aws_security_group" "bad_example" {
+  egress {
+    cidr_blocks = ["0.0.0.0/0"]
+  }
+}
+
+ resource "aws_security_group_rule" "bad_example" {
+  type = "ingress"
+  cidr_blocks = ["0.0.0.0/0"]
+}
+
+"""
+
+_IAC_NO_VULNERABILITIES = """
+resource "aws_network_acl_rule" "bad_example" {
+  egress         = false
+  protocol       = "tcp"
+  from_port      = 22
+  to_port        = 22
+  rule_action    = "allow"
+  cidr_block     = "12.13.14.15"
+}
+"""
+
+
 @pytest.fixture(autouse=True)
 def do_not_use_real_config_dir(monkeypatch, tmp_path):
     """
