+++ tests/functional/iac/test_iac_scan_diff.py
@@ -2,9 +2,9 @@ from pathlib import Path
 
 import pytest
 
+from tests.conftest import _IAC_SINGLE_VULNERABILITY
 from tests.functional.utils import run_ggshield_iac_scan
 from tests.repository import Repository
-from tests.unit.conftest import _IAC_SINGLE_VULNERABILITY
 
 
 def test_iac_scan_diff_unchanged(tmp_path: Path) -> None:
