+++ tests/unit/iac/output/test_iac_json_output.py
@@ -8,12 +8,12 @@ from click.testing import CliRunner, Result
 from pytest_voluptuous import S
 
 from ggshield.cmd.main import cli
-from tests.unit.conftest import (
+from tests.conftest import (
     _IAC_MULTIPLE_VULNERABILITIES,
     _IAC_NO_VULNERABILITIES,
     _IAC_SINGLE_VULNERABILITY,
-    my_vcr,
 )
+from tests.unit.conftest import my_vcr
 
 
 INCIDENT_SCHEMA = validators.Schema(
