+++ tests/unit/iac/output/test_iac_text_output.py
@@ -5,14 +5,12 @@ from click.testing import CliRunner, Result
 
 from ggshield.cmd.main import cli
 from ggshield.core.errors import ExitCode
-from tests.unit.conftest import (
+from tests.conftest import (
     _IAC_MULTIPLE_VULNERABILITIES,
     _IAC_NO_VULNERABILITIES,
     _IAC_SINGLE_VULNERABILITY,
-    assert_invoke_exited_with,
-    assert_invoke_ok,
-    my_vcr,
 )
+from tests.unit.conftest import assert_invoke_exited_with, assert_invoke_ok, my_vcr
 
 
 @my_vcr.use_cassette("test_iac_scan_single_vulnerability")
